Code optimizing and clean-up with using static analyzers - many data types are changed - added casts where needed - focus to double type instead float

+++ Engine/common_features/episode_state.h
@@ -5,86 +5,86 @@
 
 class LvlExit
 {
-public:
-    enum exitLevelCodes
-    {
-        EXIT_ReplayRequest = -5,
-        EXIT_MenuExit = -3,
-        EXIT_Error = -2,
-        EXIT_PlayerDeath = -1,
-        EXIT_Closed = -4,
-        EXIT_Neutral = 0,
-        EXIT_Card = 1,
-        EXIT_Ball = 2,
-        EXIT_OffScreen = 3,
-        EXIT_Key = 4,
-        EXIT_Crystal = 5,
-        EXIT_Warp = 6,
-        EXIT_Star = 7,
-        EXIT_Tape = 8
-    };
+    public:
+        enum exitLevelCodes
+        {
+            EXIT_ReplayRequest = -5,
+            EXIT_MenuExit = -3,
+            EXIT_Error = -2,
+            EXIT_PlayerDeath = -1,
+            EXIT_Closed = -4,
+            EXIT_Neutral = 0,
+            EXIT_Card = 1,
+            EXIT_Ball = 2,
+            EXIT_OffScreen = 3,
+            EXIT_Key = 4,
+            EXIT_Crystal = 5,
+            EXIT_Warp = 6,
+            EXIT_Star = 7,
+            EXIT_Tape = 8
+        };
 };
 
 class WldExit
 {
-public:
-    enum ExitWorldCodes
-    {
-        EXIT_close = -2,
-        EXIT_error = -1,
-        EXIT_exitWithSave = 1,
-        EXIT_exitNoSave = 2,
-        EXIT_beginLevel = 0
-    };
+    public:
+        enum ExitWorldCodes
+        {
+            EXIT_close = -2,
+            EXIT_error = -1,
+            EXIT_exitWithSave = 1,
+            EXIT_exitNoSave = 2,
+            EXIT_beginLevel = 0
+        };
 };
 
 struct PlayerState
 {
     PlayerState() : characterID(1), stateID(1) {}
-    int characterID;
-    int stateID;
+    unsigned long characterID;
+    unsigned long stateID;
     saveCharState _chsetup;
 };
 
 class EpisodeState
 {
-public:
-    EpisodeState();
-    ~EpisodeState();
-    void reset();//!< Sets initial state of episode
-    bool load();
-    bool save();
-    int numOfPlayers;//!< Number of players
-    bool episodeIsStarted;
-    bool isEpisode;
-    bool isHubLevel;
-    bool isTestingModeL;
-    bool isTestingModeW;
-    GamesaveData game_state;
+    public:
+        EpisodeState();
+        ~EpisodeState();
+        void reset();//!< Sets initial state of episode
+        bool load();
+        bool save();
+        int numOfPlayers;//!< Number of players
+        bool episodeIsStarted;
+        bool isEpisode;
+        bool isHubLevel;
+        bool isTestingModeL;
+        bool isTestingModeW;
+        GamesaveData game_state;
 
-    PlayerState getPlayerState(int playerID);
-    void setPlayerState(int playerID, PlayerState &state);
+        PlayerState getPlayerState(int playerID);
+        void setPlayerState(int playerID, PlayerState &state);
 
-    QString WorldFile;
-    QString WorldPath;
-    QString saveFileName;
-    QString _episodePath;
-    int _recent_ExitCode_level;
-    int _recent_ExitCode_world;
+        QString WorldFile;
+        QString WorldPath;
+        QString saveFileName;
+        QString _episodePath;
+        int _recent_ExitCode_level;
+        int _recent_ExitCode_world;
 
-    QString LevelFile;
-    QString LevelFile_hub;
-    QString LevelPath;
-        int LevelTargetWarp;
-    int gameType;
-    enum gameTypes
-    {
-        Testing=0,
-        Episode=1,
-        Battle,
-        Race
-    };
-    bool replay_on_fail;
+        QString LevelFile;
+        QString LevelFile_hub;
+        QString LevelPath;
+        unsigned long LevelTargetWarp;
+        int gameType;
+        enum gameTypes
+        {
+            Testing = 0,
+            Episode = 1,
+            Battle,
+            Race
+        };
+        bool replay_on_fail;
 };
 
 #endif // EPISODESTATE_H
