+++ Engine/data_configs/obj_bgo.h
@@ -26,10 +26,10 @@ struct obj_bgo
 {
     /*   OpenGL    */
     bool isInit;
-    PGE_Texture * image;
+    PGE_Texture *image;
     GLuint textureID;
-    long textureArrayId;
-    long animator_ID;
+    int textureArrayId;
+    int animator_ID;
     /*   OpenGL    */
 
     BgoSetup setup;
